+++ config.js
@@ -1,5 +1,5 @@
 window.VERDEAI_CONFIG = {
-  version: "8.3",
+  version: "8.4",
   apiBaseUrl: "http://localhost:8080",
   useBackend: false,
   staticMode: true